Moved files to better folder structure. Moved code from the Monolithic main.js to Game, Scene, MapScene, FightScene, and Map classes.
