+++ public/scripts/Game.js
@@ -1,0 +1,56 @@
+import {Player} from './world/Character.js';
+import {MapScene} from './world/MapScene.js';
+import {FightScene} from './fight/FightScene.js';
+
+export class Game {
+  static initialize() {
+    const game = new Game();
+    Game.instance = game;
+  }
+
+  static getInstance() {
+    return Game.instance;
+  }
+
+  constructor(data) {
+    this.stage = new Konva.Stage({
+      container: 'container',
+      width: 1000,
+      height: 500,
+    });
+
+    this.player = new Player({
+      x: 40,
+      y: this.stage.height() - 80,
+      width: 40,
+      height: 40,
+      image: 'assets/bobby.jpg',
+      colour: 'grey',
+      hp: 100,
+    });
+
+    const container = this.stage.container();
+    container.tabIndex = 1;
+    container.focus();
+  }
+
+  start() {
+    this.mapScene = new MapScene({stage: this.stage, player: this.player});
+    this.fightScene = new FightScene({stage: this.stage, player: this.player});
+
+    this.current_scene = this.mapScene;
+    this.current_scene.switchTo({stage: this.stage, player: this.player});
+  }
+
+  switchToFight() {
+    this.current_scene.switchFrom({stage: this.stage, player: this.player});
+    this.current_scene = this.fightScene;
+    this.current_scene.switchTo({stage: this.stage, player: this.player});
+  }
+
+  switchToMap() {
+    this.current_scene.switchFrom({stage: this.stage, player: this.player});
+    this.current_scene = this.mapScene;
+    this.current_scene.switchTo({stage: this.stage, player: this.player});
+  }
+}
