+++ public/scripts/Scene.js
@@ -1,0 +1,13 @@
+export class Scene {
+  constructor(data) {
+
+  }
+
+  switchTo() {
+    console.log('Scene switchTo not implemented!');
+  }
+
+  switchFrom() {
+    console.log('Scene switchFrom not implemented!');
+  }
+}
