+++ public/scripts/util/helper_functions.js
