+++ public/scripts/world/Wall.js
